feat: generic attachments for messages components
